+++ src/main/kotlin/com/cinephiliapaymentservice/domain/Movie.kt
@@ -1,0 +1,8 @@
+package com.cinephiliapaymentservice.domain
+
+import java.util.*
+
+data class Movie(
+    var id: UUID,
+    var price: Double
+)
