+++ src/main/kotlin/com/cinephiliapaymentservice/domain/User.kt
@@ -1,0 +1,7 @@
+package com.cinephiliapaymentservice.domain
+
+import java.util.*
+
+data class User(
+    val id: UUID
+)
